+++ src/addons/ui_skybrush_studio.py
@@ -41,7 +41,10 @@ for candidate in candidates:
 #############################################################################
 # imports needed by the addon
 
-from sbstudio.plugin.lists import SKYBRUSH_UL_lightfxlist
+from sbstudio.plugin.lists import (
+    SKYBRUSH_UL_lightfxlist,
+    SKYBRUSH_UL_scheduleoverridelist,
+)
 from sbstudio.plugin.menus import GenerateMarkersMenu
 from sbstudio.plugin.model import (
     DroneShowAddonFileSpecificSettings,
@@ -53,6 +56,7 @@ from sbstudio.plugin.model import (
     LightEffect,
     LightEffectCollection,
     SafetyCheckProperties,
+    ScheduleOverride,
     StoryboardEntry,
     Storyboard,
 )
@@ -62,6 +66,7 @@ from sbstudio.plugin.operators import (
     AppendFormationToStoryboardOperator,
     ApplyColorsToSelectedDronesOperator,
     CreateFormationOperator,
+    CreateNewScheduleOverrideEntryOperator,
     CreateNewStoryboardEntryOperator,
     CreateLightEffectOperator,
     CreateTakeoffGridOperator,
@@ -80,6 +85,7 @@ from sbstudio.plugin.operators import (
     RecalculateTransitionsOperator,
     RemoveFormationOperator,
     RemoveLightEffectOperator,
+    RemoveScheduleOverrideEntryOperator,
     RemoveStoryboardEntryOperator,
     ReorderFormationMarkersOperator,
     ReturnToHomeOperator,
@@ -139,6 +145,7 @@ types = (
     FormationsPanelProperties,
     LightEffect,
     LightEffectCollection,
+    ScheduleOverride,
     StoryboardEntry,
     Storyboard,
     LEDControlPanelProperties,
@@ -165,6 +172,8 @@ operators = (
     MoveStoryboardEntryUpOperator,
     SelectStoryboardEntryForCurrentFrameOperator,
     RemoveStoryboardEntryOperator,
+    CreateNewScheduleOverrideEntryOperator,
+    RemoveScheduleOverrideEntryOperator,
     UpdateFrameRangeFromStoryboardOperator,
     UpdateTimeMarkersFromStoryboardOperator,
     CreateLightEffectOperator,
@@ -194,7 +203,7 @@ operators = (
 )
 
 #: List widgets in this addon.
-lists = (SKYBRUSH_UL_lightfxlist,)
+lists = (SKYBRUSH_UL_lightfxlist, SKYBRUSH_UL_scheduleoverridelist)
 
 #: Menus in this addon
 menus = (GenerateMarkersMenu,)