+++ src/modules/sbstudio/plugin/lists/__init__.py
@@ -1,3 +1,4 @@
 from .light_effects import SKYBRUSH_UL_lightfxlist
+from .schedule_overrides import SKYBRUSH_UL_scheduleoverridelist
 
-__all__ = ("SKYBRUSH_UL_lightfxlist",)
+__all__ = ("SKYBRUSH_UL_lightfxlist", "SKYBRUSH_UL_scheduleoverridelist")