+++ src/modules/sbstudio/plugin/model/__init__.py
@@ -6,7 +6,7 @@ from .object_props import DroneShowAddonObjectProperties
 from .safety_check import SafetyCheckProperties
 from .settings import DroneShowAddonFileSpecificSettings
 from .show import DroneShowAddonProperties
-from .storyboard import StoryboardEntry, Storyboard
+from .storyboard import ScheduleOverride, StoryboardEntry, Storyboard
 
 __all__ = (
     "DroneShowAddonFileSpecificSettings",
@@ -18,6 +18,7 @@ __all__ = (
     "LightEffect",
     "LightEffectCollection",
     "SafetyCheckProperties",
+    "ScheduleOverride",
     "StoryboardEntry",
     "Storyboard",
 )