+++ src/modules/sbstudio/plugin/operators/recalculate_transitions.py
@@ -265,6 +265,55 @@ def calculate_mapping_for_transition_into_storyboard_entry(
     return result
 
 
+def calculate_departure_index_of_drone(
+    drone,
+    drone_index: int,
+    previous_entry: StoryboardEntry,
+    previous_entry_index: int,
+    previous_mapping: Optional[Mapping],
+    objects_in_previous_formation,
+) -> int:
+    """Calculates the departure index of a drone (i.e. the index of the source
+    marker that a drone is associated to) in a transition.
+    """
+    # In the departure sequence, the index of the drone is dictated
+    # by the index of its associated marker / object within the
+    # previous formation.
+    if previous_mapping:
+        previous_target_index = previous_mapping[drone_index]
+        if previous_target_index is None:
+            # drone did not participate in the previous formation
+            return 0
+        else:
+            return previous_target_index
+    else:
+        # Previous mapping not known. All is not lost, however; we
+        # can find which point in the previous formation the drone
+        # must have belonged to by finding the constraint that binds
+        # the drone to the previous storyboard entry
+        if previous_entry is not None:
+            previous_constraint = find_transition_constraint_between(
+                drone=drone, storyboard_entry=previous_entry
+            )
+            if previous_constraint is not None:
+                previous_obj = previous_constraint.target
+            else:
+                # Either the drone did not participate in the previous formation,
+                # or the previous storyboard entry is the first one in the
+                # storyboard, in which case there are no constraints
+                if previous_entry_index == 0:
+                    return drone_index
+
+                previous_obj = None
+            return objects_in_previous_formation.find(previous_obj)
+        else:
+            # This is the first entry. If we are calculating a
+            # transition _into_ the first entry, the drones are
+            # simply ordered according to how they are placed in the
+            # Drones collection
+            return drone_index
+
+
 def update_transition_constraint_properties(drone, entry: StoryboardEntry, marker, obj):
     """Updates the constraint that attaches a drone to its target in a
     transition.
@@ -364,6 +413,7 @@ def update_transition_constraint_influence(
 
 def update_transition_for_storyboard_entry(
     entry: StoryboardEntry,
+    entry_index: int,
     drones,
     *,
     get_positions_of,
@@ -377,6 +427,8 @@ def update_transition_for_storyboard_entry(
 
     Parameters:
         entry: the storyboard entry
+        entry_index: index of the storyboard entry
+        drones: the drones in the scene
         previous_entry: the storyboard entry that precedes the given entry;
             `None` if the given entry is the first one
         previous_mapping: the mapping from drone indices to target point
@@ -384,7 +436,6 @@ def update_transition_for_storyboard_entry(
             not known or if the given entry is the first one. Used for
             staggered transitions to determine when a given drone should
             depart from the previous formation
-        drones: the drones in the scene
         start_of_scene: the first frame of the scene
         start_of_next: the frame where the _next_ storyboard entry starts;
             `None` if this is the last storyboard entry
@@ -428,6 +479,10 @@ def update_transition_for_storyboard_entry(
     objects_in_formation = _LazyFormationObjectList(entry)
     objects_in_previous_formation = _LazyFormationObjectList(previous_entry)
 
+    # Create a mapping that maps indices of points in the source formation to
+    # the corresponding schedule overrides (if any)
+    schedule_override_map = entry.get_enabled_schedule_override_map()
+
     # Now we have the index of the target point that each drone
     # should be mapped to, and we have `None` for those drones that
     # will not participate in the formation
@@ -447,10 +502,21 @@ def update_transition_for_storyboard_entry(
 
             windup_start_frame = end_of_previous
             start_frame = entry.frame_start
+            departure_delay = 0
+            arrival_delay = 0
+            departure_index: Optional[int] = None
 
             if entry.is_staggered:
                 # Determine the index of the drone in the departure sequence
                 # and in the arrival sequence
+                departure_index = calculate_departure_index_of_drone(
+                    drone,
+                    drone_index,
+                    previous_entry,
+                    entry_index - 1,
+                    previous_mapping,
+                    objects_in_previous_formation,
+                )
 
                 # In the departure sequence, the index of the drone is dictated
                 # by the index of its associated marker / object within the
@@ -495,16 +561,36 @@ def update_transition_for_storyboard_entry(
                     num_drones_transitioning - arrival_index - 1
                 )
 
-                windup_start_frame += departure_delay
-                start_frame += arrival_delay
-
-                if windup_start_frame >= start_frame:
-                    raise SkybrushStudioError(
-                        f"Not enough time to plan staggered transition to "
-                        f"formation {entry.name!r} at drone index {drone_index+1} "
-                        f"(1-based). Try decreasing departure or arrival delay "
-                        f"or allow more time for the transition."
+            if schedule_override_map:
+                # Determine the index of the drone in the departure sequence
+                # so we can look up whether there is an override for it. Note
+                # that we do not need to do this again if we already have the
+                # departure index
+                if departure_index is None:
+                    departure_index = calculate_departure_index_of_drone(
+                        drone,
+                        drone_index,
+                        previous_entry,
+                        entry_index - 1,
+                        previous_mapping,
+                        objects_in_previous_formation,
                     )
+
+                override = schedule_override_map.get(departure_index)
+                if override:
+                    departure_delay = override.pre_delay
+                    arrival_delay = -override.post_delay
+
+            windup_start_frame += departure_delay
+            start_frame += arrival_delay
+
+            if windup_start_frame >= start_frame:
+                raise SkybrushStudioError(
+                    f"Not enough time to plan staggered transition to "
+                    f"formation {entry.name!r} at drone index {drone_index+1} "
+                    f"(1-based). Try decreasing departure or arrival delay "
+                    f"or allow more time for the transition."
+                )
 
             # start_frame can be earlier than entry.frame_start for
             # staggered arrivals.
@@ -541,6 +627,9 @@ class RecalculationTask:
     entry: StoryboardEntry
     """The _target_ entry of the transition to recalculate."""
 
+    entry_index: int
+    """Index of the target entry of the transition."""
+
     previous_entry: Optional[StoryboardEntry] = None
     """The entry that precedes the target entry in the storyboard; `None` if the
     target entry is the first one.
@@ -555,6 +644,7 @@ class RecalculationTask:
     def for_entry_by_index(cls, entries: Sequence[StoryboardEntry], index: int):
         return cls(
             entries[index],
+            index,
             entries[index - 1] if index > 0 else None,
             entries[index + 1].frame_start if index + 1 < len(entries) else None,
         )
@@ -582,6 +672,7 @@ def recalculate_transitions(
         for task in tasks:
             previous_mapping = update_transition_for_storyboard_entry(
                 task.entry,
+                task.entry_index,
                 drones,
                 get_positions_of=get_positions_of,
                 previous_entry=task.previous_entry,